requirements.txt
# ==================== CORE DEPENDENCIES ====================
streamlit>=1.28.0
Pillow>=10.0.0
requests>=2.28.0

# ==================== AI & ML ====================
google-generativeai>=0.3.0
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0

# ==================== VOICE & AUDIO ====================
SpeechRecognition>=3.10.0
gTTS>=2.4.0
pydub>=0.25.1
audio-recorder-streamlit>=0.0.8

# ==================== FIREBASE AUTHENTICATION ====================
# Using REST API - no additional package needed
# Firebase auth handled via requests library

# ==================== UI COMPONENTS ====================
# All UI handled by Streamlit native components

# ==================== UTILITIES ====================
python-dateutil>=2.8.2
pytz>=2023.3
